Major changes:
Messed around with the default presets values (Still needs some work, specially in the case where we are transitioning from deep night into morning)

VEM now supports default presets to be replaced for another registered dynamic preset.

Refactor/Stability, changes in the time module
There was a bug that allowed old presets from a past map to still be in the lists of the sorted presets when a new map begun.

Now we have a vanilla preset placeholder that gets filled up getting the default values while we load the presets on level load. The Vanilla preset will be loaded as a default until the user indicates via chat which preset does he want or when he enables the time cycle.

Added utility functions for tables.

TODO : We need to find a better way of iterating through the presets when the time cycle is enabled, as it is right now we are iterating through ALL of the sorted presets causing the presets, that aren't the current nor the next, to be initiliazed and destroyed in a matter of seconds for nothing.
Haven't seen any performance issue TBH but it would be nice to do this in another way (couldn't find out a better one yet)

this commit is way to long :S

Refactored the VEM server script to be similar to the client one as a separated file and now the chat commands work properly.

Patches seems to be working fine, brought some stuff from Darkness Unleashed regarding spotlights and such.