Damn... Standarized most of the day night cycle presets per map according to specific necesities for some and just default values for some others.
Now there are specific settings per map to allow a day duration and a starting hour for each.

Fixed something in the VEHandler module, I think we don't want it to pulse when we are "fading into"

Multiple attempts to fix water issues for night time, most of them seem have worked :) BUT FUCK YOU XP4_EARTHQUAKE., fuck you and your damn pond in US HQ.

Of course the giant if in VEManager Client onLevelLoaded can be handled better and shorter... but I'm just don't care at this point.